+++ src/Lykke.Service.Bil2IndexerWebApi/Services/IBlockQueryFacade.cs
@@ -8,7 +8,7 @@ namespace Lykke.Service.Bil2IndexerWebApi.Services
     public interface IBlockQueryFacade
     {
         Task<BlockModel> GetBlockByIdOrDefault(string blockchainType, BlockId id);
-        Task<BlockModel> GetBlockByNumberOrDefault(string blockchainType, int number);
+        Task<BlockModel> GetBlockByNumberOrDefault(string blockchainType, long number);
         Task<IReadOnlyCollection<BlockModel>> GetBlocks(string blockchainType, int limit, bool orderAsc, long? startingAfter, long? endingBefore);
         Task<BlockModel> GetLastIrreversibleBlockAsync(string blockchainType);
         Task<BlockModel> GetLastBlockAsync(string blockchainType);
